Possibly added CustomColor and CustomText support
Untested, and there is a lot of other things that I don't fully understand yet with this library. One such thing is that I don't really know how to do some of the things I used to do with the old library, such as follow and create pointers. In this library, it seems a bit harder to create pointers. I also don't know if the objects I am mutating are going to persist in the assets files (they should, but I don't know for sure)